Shockwave Merge Resolution
The shockwave attack (not the particle FX) had it's dimensions changed in the last update (specifically, it is now 4500 units long and 700 units wide). The current shockwave FX does not reflect this.

The NS_TEMPPlayerShockwave system is based on the particle effect before Laura's last update to it, but it should resize itself correctly. Feel free to swap to this effect if the Shockwave particle size is confusing.